+++ ProjectSettings/EditorBuildSettings.asset
@@ -7,3 +7,4 @@ EditorBuildSettings:
   m_Scenes: []
   m_configObjects:
     com.unity.input.settings: {fileID: 11400000, guid: 5eb1441859adc18458b4b7ba206f2980, type: 2}
+    xr.sdk.mock-hmd.settings: {fileID: 11400000, guid: 4b46ab7a81ef4e04e88c26bb6ad6f136, type: 2}
